+++ app/src/components/helpers/TransferAdvanceOption/TransferAdvanceOption.scss
@@ -1,0 +1,17 @@
+@import '../../../scss/nav-selector.mixin';
+@import '../../../scss/card-tooltip.mixin';
+
+.advance-option {
+  .form-selector {
+    @include navSelector();
+    display: flex;
+  }
+
+  ::v-deep .v-messages {
+    text-align: right;
+  }
+}
+
+.v-tooltip__content {
+  @include cardTooltip();
+}
